Refactor around direct client
