+++ src/routers/direct-api.ts
@@ -3,12 +3,12 @@ import * as auth from '../auth';
 
 const router = Router();
 
-router.get('/domains', auth.findClient, async (req, res) => {
+router.get('/domains', auth.findClient, async (req: auth.CheckedRequest, res) => {
   const domains = await req.client.getDomains();
   res.send(domains);
 });
 
-router.get('/domains/:domainId/talks', auth.findClient, async (req, res) => {
+router.get('/domains/:domainId/talks', auth.findClient, async (req: auth.CheckedRequest, res) => {
   const talks = await req.client.getTalks(req.params.domainId);
   res.send(talks);
 });
@@ -18,11 +18,11 @@ router.get('/domains/:domainId/talks/:talkId/messages', (req, res) => {
   res.send('');
 });
 
-router.post('/domains/:domainId/talks/:talkId/messages', auth.findClient, async (req, res) => {
+router.post('/domains/:domainId/talks/:talkId/messages', auth.findClient, async (req: auth.CheckedRequest, res) => {
   const q = req.params;
   const result = await req.client.sendTextMessage(q.domainId, q.talkId, req.body)
-    .then((res: string) => ({status: 200, body: res}))
-    .catch((err: string) => {
+    .then((res) => ({status: 200, body: res}))
+    .catch((err) => {
       if (err === 'NotFound') {
         return {status: 404, body: ''};
       } else {
