--- src/types/express-custom.d.ts
@@ -1,7 +1,0 @@
-/// <reference types="express" />
-
-declare namespace Express {
-  export interface Request {
-    client?: any  // TOOD: DirectClient
-  }
-}
